Implemented the Create Workspace for a particular user feature + minor optimizations

+++ ToDoSpringboot/src/main/java/com/jonathan/todo/repository/UserRepository.java
@@ -6,8 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.jonathan.todo.model.User;
 
-public interface UserRepository extends JpaRepository<User, Integer>{
-	Optional<User> findByUsername(String username);
+public interface UserRepository extends JpaRepository<User, Long>{
+	Optional<User> findByUsername(String username); // check if this and below are mandatory
 	
 	boolean existsByUsername(String username);
 }
